+++ tls-awslc/src/macosTest/kotlin/io/github/fukusaka/keel/tls/awslc/TestEngine.macos.kt
@@ -1,0 +1,7 @@
+package io.github.fukusaka.keel.tls.awslc
+
+import io.github.fukusaka.keel.core.IoEngine
+import io.github.fukusaka.keel.core.IoEngineConfig
+import io.github.fukusaka.keel.engine.kqueue.KqueueEngine
+
+actual fun createTestEngine(): IoEngine = KqueueEngine(IoEngineConfig(threads = 1))